+++ e2e/nx-firebase-e2e/test-utils/index.ts
@@ -1,6 +1,6 @@
 export * from './test-helpers'
 export * from './test-logger'
-export * from './utils'
+export * from './test-imports'
 export * from './test-commands'
 export * from './test-project-data'
 export * from './test-apps'
